[refactor] pipelines.ts model 주석 처리 후 경로 변경
