+++ frontend/packages/dev-console/src/components/pipelineruns/PipelineRunsPage.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { getBadgeFromType } from '@console/shared';
-import { PipelineRunModel } from '../../models';
+import { PipelineRunModel } from '../../../../../public/models';
 import ProjectListPage from '../projects/ProjectListPage';
 import PipelineRunsResourceList from './PipelineRunsResourceList';
