+++ frontend/packages/dev-console/src/components/pipelineruns/PipelineRunsResourceList.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { getBadgeFromType } from '@console/shared';
 import { ListPage } from '@console/internal/components/factory';
 import { referenceForModel } from '@console/internal/module/k8s';
-import { PipelineRunModel } from '../../models';
+import { PipelineRunModel } from '../../../../../public/models';
 import { runFilters } from '../pipelines/detail-page-tabs/PipelineRuns';
 import PipelineRunsList from './list-page/PipelineRunList';
