+++ frontend/packages/dev-console/src/components/pipelines/PipelinesPage.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { getBadgeFromType } from '@console/shared';
 import { withStartGuide } from '@console/internal/components/start-guide';
-import { PipelineModel } from '../../models';
+import { PipelineModel } from '../../../../../public/models';
 import ProjectListPage from '../projects/ProjectListPage';
 import PipelinesResourceList from './PipelinesResourceList';
