+++ frontend/packages/dev-console/src/components/pipelines/PipelinesResourceList.tsx
@@ -3,7 +3,7 @@ import { getBadgeFromType } from '@console/shared';
 import { referenceForModel } from '@console/internal/module/k8s';
 import { Firehose } from '@console/internal/components/utils';
 import { FireMan_ as FireMan } from '@console/internal/components/factory';
-import { PipelineModel } from '../../models';
+import { PipelineModel } from '../../../../../public/models';
 import PipelineAugmentRunsWrapper from './list-page/PipelineAugmentRunsWrapper';
 import { filters } from './list-page/PipelineAugmentRuns';
