+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/PipelineRuns.tsx
@@ -7,7 +7,7 @@ import {
   pipelineRunStatusFilter,
 } from '../../../utils/pipeline-filter-reducer';
 import { ListFilterId, ListFilterLabels } from '../../../utils/pipeline-utils';
-import { PipelineRunModel } from '../../../models';
+import { PipelineRunModel } from '../../../../../../public/models';
 
 export const runFilters = [
   {
