+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/TriggerTemplateDetails.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { SectionHeading, ResourceSummary } from '@console/internal/components/utils';
-import { EventListenerModel, PipelineModel } from '../../../models';
+import { EventListenerModel } from '../../../models';
+import { PipelineModel } from '../../../../../../public/models'
 import ResourceLinkList from '../resource-overview/ResourceLinkList';
 import {
   useTriggerTemplateEventListenerNames,
