+++ frontend/packages/dev-console/src/components/pipelines/list-page/PipelineAugmentRunsWrapper.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { ListPageWrapper_ as ListPageWrapper } from '@console/internal/components/factory';
 import { Firehose } from '@console/internal/components/utils';
 import { Resource, getResources } from '../../../utils/pipeline-augment';
-import { PipelineModel } from '../../../models';
+import { PipelineModel } from '../../../../../../public/models';
 import PipelineAugmentRuns, { filters } from './PipelineAugmentRuns';
 import PipelineList from './PipelineList';
