+++ frontend/packages/dev-console/src/components/pipelines/list-page/PipelineList.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { SortByDirection } from '@patternfly/react-table';
 import { Table } from '@console/internal/components/factory';
 import { PropPipelineData } from '../../../utils/pipeline-augment';
-import { PipelineModel } from '../../../models';
+import { PipelineModel } from '../../../../../../public/models';
 import PipelineHeader from './PipelineHeader';
 import PipelineRow from './PipelineRow';
