+++ frontend/packages/dev-console/src/components/pipelines/modals/common/PipelineResourceDropdown.tsx
@@ -8,7 +8,7 @@ import {
   WatchK8sResource,
 } from '@console/internal/components/utils/k8s-watch-hook';
 import { LoadingInline } from '@console/internal/components/utils';
-import { PipelineResourceModel } from '../../../../models';
+import { PipelineResourceModel } from '../../../../../../../public/models';
 import { PipelineResourceKind } from '../../../../utils/pipeline-augment';
 import { CREATE_PIPELINE_RESOURCE } from './const';
 import './PipelineResourceDropdown.scss';
