+++ frontend/packages/dev-console/src/components/pipelines/modals/common/utils.ts
@@ -9,7 +9,7 @@ import {
   PipelineRunReferenceResource,
   PipelineRunResource,
 } from '../../../../utils/pipeline-augment';
-import { PipelineRunModel } from '../../../../models';
+import { PipelineRunModel } from '../../../../../../../public/models';
 import { getPipelineRunParams, getPipelineRunWorkspaces } from '../../../../utils/pipeline-utils';
 import { CREATE_PIPELINE_RESOURCE, initialResourceFormValues } from './const';
 import { CommonPipelineModalFormikValues, PipelineModalFormResource } from './types';
