+++ frontend/packages/dev-console/src/components/taskruns/TaskRunDetails.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { SectionHeading, ResourceSummary } from '@console/internal/components/utils';
 import { K8sResourceKind } from '@console/internal/module/k8s';
-import { TaskRunModel } from '../../models';
+import { TaskRunModel } from '../../../../../public/models';
 
 export interface TaskRunDetailsProps {
   obj: K8sResourceKind;
