+++ frontend/packages/dev-console/src/plugin.tsx
@@ -18,6 +18,7 @@ import {
 } from '@console/internal/models';
 // import * as helmIcon from '@console/internal/imgs/logos/helm.svg';
 import * as models from './models';
+import * as hypercloudModels from '../../../public/models';
 import { getKebabActionsForKind } from './utils/kebab-actions';
 import { tknPipelineAndPipelineRunsResources, getPipelinesAndPipelineRunsForResource } from './utils/pipeline-plugin-utils';
 import { FLAG_OPENSHIFT_PIPELINE, ALLOW_SERVICE_BINDING } from './const';
@@ -29,7 +30,8 @@ import * as yamlIcon from './images/yaml.svg';
 // import * as dockerfileIcon from './images/dockerfile.svg';
 import * as pipelineIcon from './images/pipeline.svg';
 
-const { ClusterTaskModel, PipelineModel, PipelineResourceModel, PipelineRunModel, TaskModel, TaskRunModel, EventListenerModel, TriggerTemplateModel, TriggerBindingModel } = models;
+const { EventListenerModel, TriggerTemplateModel, TriggerBindingModel } = models;
+const { ClusterTaskModel, PipelineModel, PipelineResourceModel, PipelineRunModel, TaskModel, TaskRunModel } = hypercloudModels;
 
 type ConsumedExtensions = ModelDefinition | ModelFeatureFlag | HrefNavItem | ResourceClusterNavItem | SeparatorNavItem | ResourceNSNavItem | ResourceListPage | ResourceDetailsPage | Perspective | RoutePage | ReduxReducer | KebabActions | OverviewResourceTab | OverviewCRD | YAMLTemplate | OverviewTabSection | AddAction;
 
@@ -43,7 +45,7 @@ const plugin: Plugin<ConsumedExtensions> = [
   {
     type: 'FeatureFlag/Model',
     properties: {
-      model: models.PipelineModel,
+      model: hypercloudModels.PipelineModel,
       flag: FLAG_OPENSHIFT_PIPELINE,
     },
   },
